refactor - remoção de arquivos desnecessários, estilização e fix do tailwind
