+++ package.json
@@ -16,7 +16,9 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
+    "@tailwindcss/forms": "^0.5.10",
     "@tailwindcss/postcss": "^4.1.12",
+    "@tailwindcss/typography": "^0.5.16",
     "@types/axios": "^0.9.36",
     "@types/react": "^19.1.10",
     "@types/react-dom": "^19.1.7",
