+++ src/components/RadioGroup.tsx
@@ -10,20 +10,28 @@ interface RadioGroupProps {
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
-export default function RadioGroup({ name, options, selected, onChange }: RadioGroupProps) {
+export default function RadioGroup({
+  name,
+  options,
+  selected,
+  onChange,
+}: RadioGroupProps) {
   return (
-    <div className="flex space-x-4">
+    <div className="flex space-x-6">
       {options.map((opt) => (
-        <label key={opt.value} className="flex items-center space-x-2">
+        <label
+          key={opt.value}
+          className="flex items-center space-x-3 cursor-pointer"
+        >
           <input
             type="radio"
             name={name}
             value={opt.value}
             checked={selected === opt.value}
             onChange={onChange}
-            className="form-radio"
+            className="w-4 h-4 text-amber-600 bg-gray-100 border-gray-300 focus:ring-amber-500 focus:ring-2"
           />
-          <span>{opt.label}</span>
+          <span className="text-sm font-medium text-gray-700">{opt.label}</span>
         </label>
       ))}
     </div>
