--- src/pages/Register/Register.tsx
@@ -1,5 +1,0 @@
-import SinglePageForm from "../../components/SinglePageForm";
-
-export default function Register() {
-  return <SinglePageForm />;
-}
